add validator admin panel job trigger
