+++ validator/app/src/compute_horde_validator/validator/management/commands/debug_run_organic_job.py
@@ -84,8 +84,6 @@ class Command(BaseCommand):
             raise ValueError(f"{miner_uid=} did not announce it's ip address")
 
         miner=Miner.objects.get_or_create(hotkey=neuron.hotkey)[0]
-        print(f'{miner=}')
-
         miner_blacklisted = MinerBlacklist.objects.filter(miner=miner).exists()
         if miner_blacklisted:
             raise ValueError(f"{miner_uid=} with hotkey {neuron.hotkey} is blacklisted")
