+++ validator/app/src/compute_horde_validator/validator/models.py
@@ -62,6 +62,24 @@ class SyntheticJob(JobBase):
             models.UniqueConstraint(fields=["batch", "miner"], name="one_job_per_batch_per_miner"),
         ]
 
-
 class OrganicJob(JobBase):
     pass
+
+
+class AdminJobRequest(models.Model):
+    job_uuid = models.UUIDField(default=uuid.uuid4, unique=True)
+    miner = models.ForeignKey(Miner, on_delete=models.PROTECT)
+    timeout = models.PositiveIntegerField(default=0, help_text="timeout in seconds")
+
+    docker_image = models.CharField(max_length=255, help_text="docker image for job execution")
+    raw_script = models.TextField(blank=True, help_text="raw script to be executed")
+
+    args = models.TextField(blank=True, help_text="arguments passed to the script or docker image")
+    env = models.JSONField(blank=True, default=dict, help_text="environment variables for the job")
+
+    use_gpu = models.BooleanField(default=False, help_text="Whether to use GPU for the job")
+    input_url = models.URLField(blank=True, help_text="URL to the input data source", max_length=1000)
+    output_url = models.TextField(blank=True, help_text="URL for uploading output")
+
+    created_at = models.DateTimeField(auto_now_add=True)
+
