+++ validator/app/src/compute_horde_validator/validator/utils.py
@@ -2,6 +2,7 @@ import base64
 import datetime as dt
 import io
 import zipfile
+from functools import cache
 
 MACHINE_SPEC_GROUP_NAME = "machine_spec_sending"
 
@@ -28,3 +29,13 @@ class Timer:
         if self.timeout is None:
             raise ValueError("timeout was not specified")
         return self.timeout - self.passed_time()
+
+@cache
+def get_dummy_inline_zip_volume() -> str:
+    in_memory_output = io.BytesIO()
+    with zipfile.ZipFile(in_memory_output, 'w'):
+        pass
+    in_memory_output.seek(0)
+    zip_contents = in_memory_output.read()
+    base64_zip_contents = base64.b64encode(zip_contents)
+    return base64_zip_contents.decode()
